+++ src/builders/resid-builder/resid-builder.cpp
@@ -33,12 +33,12 @@ ReSIDBuilder::~ReSIDBuilder()
 }
 
 // Create a new sid emulation.
-unsigned int ReSIDBuilder::create(unsigned int sids)
+size_t ReSIDBuilder::create(size_t sids)
 {
     m_status = true;
 
     // Check available devices
-    unsigned int count = availDevices();
+    size_t count = availDevices();
 
     if (count && (count < sids))
         sids = count;
